requirements.txt
flask>=3.0
python-dotenv
pymysql
google-auth
google-auth-oauthlib
google-api-python-client
requests
